+++ web/js/033-utilities.js
@@ -2876,12 +2876,12 @@
 		}
 	};
 
-	Utilities.prototype.setCaption = function(title, description, tags) {
+	Utilities.prototype.setDescription = function(title, description, tags) {
 		// Replace CRLF by <br> and remove all useless <br>.
 		function formatText(text) {
 			text = text.replace(/<(\/?\w+)>\s*\n\s*<(\/?\w+)>/g, "<$1><$2>");
-			text = text.replace(/\n/g, "</p><p class='caption-text'>");
-			return "<p class='caption-text'>" + text + "</p>";
+			text = text.replace(/\n/g, "</p><p class='description-text'>");
+			return "<p class='description-text'>" + text + "</p>";
 		}
 
 		var nullTitle = (typeof title === "undefined") || ! title;
@@ -2889,46 +2889,46 @@
 		var nullTags = (typeof tags === "undefined") || ! tags.length;
 
 		if (! nullTitle) {
-			$("#caption-title").html(formatText(title));
+			$("#description-title").html(formatText(title));
 		} else {
-			$("#caption-title").html("");
+			$("#description-title").html("");
 		}
 
 		if (! nullDescription) {
-			$("#caption-description .description").html(formatText(description));
+			$("#description-text .description").html(formatText(description));
 		} else {
-			$("#caption-description .description").html("");
+			$("#description-text .description").html("");
 		}
 		if (! nullTags) {
 			// let textualTags = "<p class='tags'> " + Utilities._t("#tags") + ": <span class='tag'>" + tags.join("</span>, <span class='tag'>") + "</span></p>";
 			let textualTags = Utilities._t("#tags") + ": <span class='tag'>" + tags.map(tag => Utilities.addTagLink(tag)).join("</span>, <span class='tag'>") + "</span>";
-			$("#caption-tags").html(textualTags);
+			$("#description-tags").html(textualTags);
 		} else {
-			$("#caption-tags").html("");
+			$("#description-tags").html("");
 		}
 
 		if (nullTitle && nullDescription && nullTags) {
-		  $("#caption").addClass("hidden");
+		  $("#description").addClass("hidden");
 		} else {
-		  $("#caption").removeClass("hidden");
+		  $("#description").removeClass("hidden");
 		}
 	};
 
-	Utilities.prototype.setCaptionPosition = function(captionType) {
-		// Size of caption varies if on album or media
+	Utilities.prototype.setDescriptionPosition = function(captionType) {
+		// Size of description varies if on album or media
 		if (captionType === 'media') {
 			var titleHeight = parseInt($(".media-box#center .title").css("height"));
 			var mediaHeight = parseInt($(".media-box#center .media-box-inner").css("height"));
-			$("#caption").css("top", titleHeight + mediaHeight * 0.7);
-			// $("#caption").css("bottom", "");
-			$("#caption").css("height", "");
+			$("#description").css("top", titleHeight + mediaHeight * 0.7);
+			// $("#description").css("bottom", "");
+			$("#description").css("height", "");
 			var selectBoxWidth = 30;
-			$("#caption").css("right", 2 * selectBoxWidth + parseInt($("#media-select-box .select-box").css("right")));
-			$("#caption").css("max-height", "");
-			$("#caption").css("max-height", mediaHeight * 0.2);
-			$("#caption-description").css("height", "");
-			$("#caption-description").css("max-height", "");
-			$("#caption-description").css("max-height", $("#caption").height() - 2 * parseInt($("#caption").css("padding-top")) - $("#caption-title").height() - $("#caption-tags").height());
+			$("#description").css("right", 2 * selectBoxWidth + parseInt($("#media-select-box .select-box").css("right")));
+			$("#description").css("max-height", "");
+			$("#description").css("max-height", mediaHeight * 0.2);
+			$("#description-text").css("height", "");
+			$("#description-text").css("max-height", "");
+			$("#description-text").css("max-height", $("#description").height() - 2 * parseInt($("#description").css("padding-top")) - $("#description-title").height() - $("#description-tags").height());
 		} else if (captionType === 'album') {
 			// var titleHeight = parseInt($("#album-view .title").css("height"));
 			// var albumTop = parseInt($("#album-view").css("top"));
@@ -2936,12 +2936,12 @@
 			var thumbsHeight = parseInt($("#thumbs").css("height"));
 			var subalbumsHeight = parseInt($("#subalbums").css("height"));
 			// TODO: How to adapt height to different platforms?
-			$("#caption").css("top", "");
-			$("#caption").css("bottom", 0);
-			$("#caption").css("height", "");
-			$("#caption").css("max-height", "");
-			$("#caption").css("height", (albumHeight + thumbsHeight + subalbumsHeight) * 0.6);
-			$("#caption").css("max-height", (albumHeight + thumbsHeight + subalbumsHeight) * 0.6);
+			$("#description").css("top", "");
+			$("#description").css("bottom", 0);
+			$("#description").css("height", "");
+			$("#description").css("max-height", "");
+			$("#description").css("height", (albumHeight + thumbsHeight + subalbumsHeight) * 0.6);
+			$("#description").css("max-height", (albumHeight + thumbsHeight + subalbumsHeight) * 0.6);
 		}
 	};