+++ web/js/035-pinch-swipe.js
@@ -229,7 +229,7 @@
 								// zoomAfterFirstPinch = currentZoom;
 								util.setPinchButtonsPosition();
 								util.correctPrevNextPosition();
-								util.setCaptionPosition('media');
+								util.setDescriptionPosition('media');
 								PinchSwipe.setPinchButtonsVisibility();
 								util.setSelectButtonPosition();
 								mediaWidth = $(mediaSelector).css("width");
@@ -337,7 +337,7 @@
 					util.correctPrevNextPosition();
 					PinchSwipe.setPinchButtonsVisibility();
 					util.setSelectButtonPosition();
-					util.setCaptionPosition('media');
+					util.setDescriptionPosition('media');
 					mediaWidth = $(mediaSelector).css("width");
 					mediaHeight = $(mediaSelector).css("height");
 					initialZoom = PinchSwipe.screenZoom();