+++ web/js/036-functions.js
@@ -427,13 +427,13 @@
 		}
 
 		if (isMapOrPopup) {
-			$("ul#right-menu li.hide-media-caption").addClass("hidden");
+			$("ul#right-menu li.hide-descriptions-and-tags").addClass("hidden");
 		} else {
-			$("ul#right-menu li.hide-media-caption").removeClass("hidden");
-			if (env.options.hide_caption)
-				$("ul#right-menu li.hide-media-caption").addClass("selected");
+			$("ul#right-menu li.hide-descriptions-and-tags").removeClass("hidden");
+			if (env.options.hide_descriptions_and_tags)
+				$("ul#right-menu li.hide-descriptions-and-tags").addClass("selected");
 			else
-				$("ul#right-menu li.hide-media-caption").removeClass("selected");
+				$("ul#right-menu li.hide-descriptions-and-tags").removeClass("selected");
 		}
 
 		if (
@@ -553,7 +553,7 @@
 
 		if (
 			$("ul#right-menu li.hide-title").hasClass("hidden") &&
-			$("ul#right-menu li.hide-media-caption").hasClass("hidden") &&
+			$("ul#right-menu li.hide-descriptions-and-tags").hasClass("hidden") &&
 			$("ul#right-menu li.media-count").hasClass("hidden") &&
 			$("ul#right-menu li.spaced").hasClass("hidden") &&
 			$("ul#right-menu li.square-album-thumbnails").hasClass("hidden") &&
@@ -1250,10 +1250,10 @@
 		else
 			$(".title").removeClass("hidden-by-option");
 
-		if (env.options.hide_caption)
-			$("#caption").addClass("hidden-by-option");
+		if (env.options.hide_descriptions_and_tags)
+			$("#description").addClass("hidden-by-option");
 		else
-			$("#caption").removeClass("hidden-by-option");
+			$("#description").removeClass("hidden-by-option");
 
 		if (env.options.hide_bottom_thumbnails && (env.currentMedia != null || env.currentAlbum.isAlbumWithOneMedia())) {
 			$("#album-view").addClass("hidden-by-option");
@@ -1465,9 +1465,9 @@
 							if (titleCookie !== null)
 								env.options.hide_title = titleCookie;
 
-							var captionCookie = Functions.getBooleanCookie("hideCaption");
-							if (captionCookie !== null)
-								env.options.hide_caption = captionCookie;
+							var descriptionsCookie = Functions.getBooleanCookie("hideDescriptionsAndTags");
+							if (descriptionsCookie !== null)
+								env.options.hide_descriptions_and_tags = descriptionsCookie;
 
 							var bottomThumbnailsCookie = Functions.getBooleanCookie("hideBottomThumbnails");
 							if (bottomThumbnailsCookie !== null)