+++ doc/Changelog.md
@@ -4,9 +4,9 @@
 
 * fix deselection shortcut and add menu entries to deselect subalbums/media
 * add a cache that avoids requesting inexistent files more than once
-* changed shortcut for hiding title, bottom thumbnails and caption: from TAB to "h"
+* changed shortcut for hiding title, bottom thumbnails and description: from TAB to "h"
 * album view: add tags to subalbums and media
-* media view: reposition caption and add tags
+* media view: reposition description and add tags
 * bug fixes
 
 ### version v4.9.9 (Dec 27, 2020)